fix: adapting subgraphs template versioning to use semver
