+++ packages/subgraph/src/registries/daoRegistry.ts
@@ -1,5 +1,5 @@
 import {DAORegistered} from '../../generated/DAORegistry/DAORegistry';
-import {DaoTemplateV1, DaoTemplateV1_2} from '../../generated/templates';
+import {DaoTemplateV1_0_0, DaoTemplateV1_2_0} from '../../generated/templates';
 import {Dao} from '../../generated/schema';
 import {dataSource} from '@graphprotocol/graph-ts';
 
@@ -21,8 +21,8 @@ export function handleDAORegistered(event: DAORegistered): void {
   entity.createdAt = event.block.timestamp;
 
   // subscribe to templates
-  DaoTemplateV1.create(event.params.dao);
-  DaoTemplateV1_2.create(event.params.dao);
+  DaoTemplateV1_0_0.create(event.params.dao);
+  DaoTemplateV1_2_0.create(event.params.dao);
 
   entity.save();
 }
